backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# AssemblyAI transcription + audio duration detection
assemblyai>=0.30.0
mutagen>=1.47.0

# Async file I/O (large audio upload streaming)
aiofiles>=23.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# FastAPI large-file upload support
python-multipart>=0.0.9

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.27.0
